EPMEDU-823: relocation of classes
