+++ feature/home/src/main/java/com/epmedu/animeal/home/data/FeedingPointRepositoryImpl.kt
@@ -1,10 +1,10 @@
-package com.epmedu.animeal.common.data.repository
+package com.epmedu.animeal.home.data
 
-import com.epmedu.animeal.common.data.enum.AnimalPriority
-import com.epmedu.animeal.common.data.enum.AnimalState
-import com.epmedu.animeal.common.data.model.FeedingPoint
 import com.epmedu.animeal.common.data.model.MapLocation
 import com.epmedu.animeal.foundation.switch.AnimalType
+import com.epmedu.animeal.home.data.model.FeedingPoint
+import com.epmedu.animeal.home.data.model.enum.AnimalPriority
+import com.epmedu.animeal.home.data.model.enum.AnimalState
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import kotlin.random.Random
@@ -14,9 +14,9 @@ class FeedingPointRepositoryImpl : FeedingPointRepository {
     private val stubData: List<FeedingPoint> = List(25) { index ->
         FeedingPoint(
             index,
-            AnimalPriority.values()[Random.nextInt(0, AnimalPriority.values().size)],
-            AnimalState.values()[Random.nextInt(0, AnimalState.values().size)],
-            AnimalType.values()[Random.nextInt(0, AnimalType.values().size)],
+            AnimalPriority.values().random(),
+            AnimalState.values().random(),
+            AnimalType.values().random(),
             Random.nextBoolean(),
             location = MapLocation(
                 Random.nextDouble(37.362023, 37.466366),
