+++ feature/home/src/main/java/com/epmedu/animeal/home/data/model/enum/AnimalState.kt
@@ -1,4 +1,4 @@
-package com.epmedu.animeal.common.data.enum
+package com.epmedu.animeal.home.data.model.enum
 
 enum class AnimalState {
     RED, GREEN, YELLOW
