+++ feature/home/src/main/java/com/epmedu/animeal/home/data/model/enum/Remoteness.kt
@@ -1,4 +1,4 @@
-package com.epmedu.animeal.common.data.enum
+package com.epmedu.animeal.home.data.model.enum
 
 enum class Remoteness {
     NEAREST, ANY, NEAREST_10_KM
