+++ feature/home/src/main/java/com/epmedu/animeal/home/presentation/HomeScreen.kt
@@ -1,9 +1,10 @@
-package com.epmedu.animeal.home
+package com.epmedu.animeal.home.presentation
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.hilt.navigation.compose.hiltViewModel
+import com.epmedu.animeal.home.presentation.viewmodel.HomeViewModel
 
 @Composable
 fun HomeScreen() {
