+++ feature/home/src/main/java/com/epmedu/animeal/home/presentation/viewmodel/HomeState.kt
@@ -1,7 +1,7 @@
-package com.epmedu.animeal.home
+package com.epmedu.animeal.home.presentation.viewmodel
 
 import com.epmedu.animeal.common.data.model.MapLocation
-import com.epmedu.animeal.model.FeedingPointUi
+import com.epmedu.animeal.home.data.model.FeedingPointUi
 import com.mapbox.maps.Style
 
 data class HomeState(
